requirements.txt
beautifulsoup4
requests
discord
pillow
asyncio
